% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Election.R
\name{Election}
\alias{Election}
\title{R6 Super Class pour donnees occupations et saturation des parkings}
\description{
utilisee ensuite par classe Occupation et Saturation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{typeElection}}{type d'election}

\item{\code{anneeElection}}{annee de l'election}

\item{\code{xtradataParameters}}{parametre de filtre pour xtradata}

\item{\code{donneesElection}}{donnees de l'election}

\item{\code{inseeElection}}{code INSEE de la commune}

\item{\code{candidat}}{avec palette de couleur}

\item{\code{palette}}{de couleurs pour les candidats}

\item{\code{resultats}}{de la requete}

\item{\code{résultats}}{globaux commune}

\item{\code{résultats}}{par LV}

\item{\code{résultats}}{par BV}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{candidat}}{avec palette de couleur}

\item{\code{palette}}{de couleurs pour les candidats}

\item{\code{resultats}}{de la requete}

\item{\code{résultats}}{globaux commune}

\item{\code{résultats}}{par LV}

\item{\code{résultats}}{par BV}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Election$new()}}
\item \href{#method-clone}{\code{Election$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new occupation object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$new(
  typeElection = NULL,
  xtradataParameters = NULL,
  anneeElection = NULL,
  donneesElection = NULL,
  inseeElection = NULL,
  candidatsElection = NULL,
  couleursCandidats = NULL,
  resultatsRequete = NULL,
  resultatsGlobauxCommune = NULL,
  resultatsLV = NULL,
  resultatsBV = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{typeElection}}{typeElection}

\item{\code{anneeElection}}{rangeEnd}

\item{\code{donneesElection}}{rangeStep}

\item{\code{inseeElection}}{aggregation_unit}

\item{\code{candidatsElection}}{plageHoraire}

\item{\code{couleursCandidats}}{liste des parkings analyses}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Occupation` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
