% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{elections}
\alias{elections}
\alias{elections-package}
\title{elections}
\description{
This is an R package for accessing data on U.S. elections from the MIT
Election and Data Science Lab.
}
