\docType{data}
\name{state_ids}
\alias{state_ids}
\title{State Identifiers}
\format{A \code{dataframe} with 5 variables:
    \describe{
        \item{\code{state}}{State name.}
        \item{\code{state_postal}}{State U.S. Postal Service abbreviation, or two-letter ISO 1366 code.}
        \item{\code{state_fips}}{Numeric state FIPS code per FIPS PUB 5-2.}
        \item{\code{state_census}}{Numeric U.S. Census state code.}
        \item{\code{state_icpsr}}{Numeric ICPSR state code.}
    }
}
\usage{state_ids}
\description{
This dataset contains identifiers for the 50 states and the District of
Columbia.
}
\keyword{datasets}
