% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Election.R
\name{Election}
\alias{Election}
\title{R6 Super Class pour donnees occupations et saturation des parkings}
\description{
utilisee ensuite par classe Occupation et Saturation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{typeElection}}{type d'election}

\item{\code{anneeElection}}{annee de l'election}

\item{\code{xtradataParameters}}{parametre de filtre pour xtradata}

\item{\code{donneesElection}}{donnees de l'election}

\item{\code{inseeElection}}{code INSEE de la commune}

\item{\code{candidatsElection}}{candidat avec palette de couleur}

\item{\code{couleursCandidats}}{palette de couleurs pour les candidats}

\item{\code{resultatsGlobauxCommune}}{résultats globaux commune}

\item{\code{resultatsAbstention}}{résultats abstention commune}

\item{\code{resultatsLV}}{résultats par LV}

\item{\code{resultatsBV}}{résultats par BV}

\item{\code{resultatsAbstentionBV}}{résultats abstention BV}

\item{\code{resultatsAbstentionLV}}{résultats abstention LV}

\item{\code{cartoBV}}{fichier sf de localisation des bureaux de vote}

\item{\code{cartoLV}}{fichier sf de localisation des lieux de vote}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Election$new()}}
\item \href{#method-download_data}{\code{Election$download_data()}}
\item \href{#method-compute_resultats_globaux_commune}{\code{Election$compute_resultats_globaux_commune()}}
\item \href{#method-compute_abstention_commune}{\code{Election$compute_abstention_commune()}}
\item \href{#method-compute_resultats_par_BV}{\code{Election$compute_resultats_par_BV()}}
\item \href{#method-compute_resultats_par_LV}{\code{Election$compute_resultats_par_LV()}}
\item \href{#method-compute_abstention_par_BV}{\code{Election$compute_abstention_par_BV()}}
\item \href{#method-compute_abstention_par_LV}{\code{Election$compute_abstention_par_LV()}}
\item \href{#method-clone}{\code{Election$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new occupation object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$new(
  typeElection = NULL,
  xtradataParameters = NULL,
  anneeElection = NULL,
  donneesElection = NULL,
  inseeElection = NULL,
  candidatsElection = NULL,
  couleursCandidats = NULL,
  resultatsGlobauxCommune = NULL,
  resultatsAbstention = NULL,
  resultatsLV = NULL,
  resultatsBV = NULL,
  resultatsAbstentionBV = NULL,
  resultatsAbstentionLV = NULL,
  cartoBV = NULL,
  cartoLV = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{typeElection}}{typeElection}

\item{\code{xtradataParameters}}{xtradataParameters}

\item{\code{anneeElection}}{anneeElection}

\item{\code{donneesElection}}{donneesElection}

\item{\code{inseeElection}}{inseeElection}

\item{\code{candidatsElection}}{candidatsElection}

\item{\code{couleursCandidats}}{couleursCandidats}

\item{\code{resultatsGlobauxCommune}}{resultatsGlobauxCommune}

\item{\code{resultatsAbstention}}{resultatsAbstention}

\item{\code{resultatsLV}}{resultatsLV}

\item{\code{resultatsBV}}{resultatsBV}

\item{\code{resultatsAbstentionBV}}{resultatsAbstentionBV}

\item{\code{resultatsAbstentionLV}}{resultatsAbstentionLV}

\item{\code{cartoBV}}{cartoBV}

\item{\code{cartoLV}}{cartoLV}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Election` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-download_data}{}}}
\subsection{Method \code{download_data()}}{
Interroge le WS features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$download_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_resultats_globaux_commune"></a>}}
\if{latex}{\out{\hypertarget{method-compute_resultats_globaux_commune}{}}}
\subsection{Method \code{compute_resultats_globaux_commune()}}{
Calcule les résultats globaux pour la commune
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$compute_resultats_globaux_commune(donneesElection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{donneesElection}}{donnees issue du WS features}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_abstention_commune"></a>}}
\if{latex}{\out{\hypertarget{method-compute_abstention_commune}{}}}
\subsection{Method \code{compute_abstention_commune()}}{
Calcule les résultats d'abstention pour la commune
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$compute_abstention_commune(donneesElection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{donneesElection}}{donnees issue du WS features}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_resultats_par_BV"></a>}}
\if{latex}{\out{\hypertarget{method-compute_resultats_par_BV}{}}}
\subsection{Method \code{compute_resultats_par_BV()}}{
Calcule les résultats par BV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$compute_resultats_par_BV(donneesElection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{donneesElection}}{donnees issue du WS features}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_resultats_par_LV"></a>}}
\if{latex}{\out{\hypertarget{method-compute_resultats_par_LV}{}}}
\subsection{Method \code{compute_resultats_par_LV()}}{
Calcule les résultats par LV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$compute_resultats_par_LV(donneesElection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{donneesElection}}{donnees issue du WS features}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_abstention_par_BV"></a>}}
\if{latex}{\out{\hypertarget{method-compute_abstention_par_BV}{}}}
\subsection{Method \code{compute_abstention_par_BV()}}{
Calcule les résultats d'abstention par BV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$compute_abstention_par_BV(donneesElection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{donneesElection}}{donnees issue du WS features}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_abstention_par_LV"></a>}}
\if{latex}{\out{\hypertarget{method-compute_abstention_par_LV}{}}}
\subsection{Method \code{compute_abstention_par_LV()}}{
Calcule les résultats d'abstention par LV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$compute_abstention_par_LV(donneesElection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{donneesElection}}{donnees issue du WS features}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Election$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
